requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.1
pydantic-settings==2.5.2
chromadb==0.5.5
rank-bm25==0.2.2
openai==1.43.0
httpx<0.28
python-multipart==0.0.9
pypdf==5.0.1
python-docx==1.1.2
# Splitters avançados
langchain-text-splitters>=0.0.1,<0.2
tiktoken>=0.5.2
